+++ core/src/elements/ons-popover/index.js
@@ -544,14 +544,11 @@ class PopoverElement extends BaseElement {
 
     contentReady(this, () => {
       this._margin = this._margin || parseInt(window.getComputedStyle(this).getPropertyValue('top'));
-      this._radius = parseInt(window.getComputedStyle(this._content).getPropertyValue('border-radius'));
+      this._radius = parseInt(window.getComputedStyle(this._content).getPropertyValue('border-top-left-radius'));
 
       this._mask.addEventListener('click', this._boundCancel, false);
 
       this._resetBackButtonHandler();
-
-      this._popover.addEventListener('DOMNodeInserted', this._boundOnChange, false);
-      this._popover.addEventListener('DOMNodeRemoved', this._boundOnChange, false);
 
       window.addEventListener('resize', this._boundOnChange, false);
     });
@@ -563,9 +560,6 @@ class PopoverElement extends BaseElement {
 
       this._backButtonHandler.destroy();
       this._backButtonHandler = null;
-
-      this._popover.removeEventListener('DOMNodeInserted', this._boundOnChange, false);
-      this._popover.removeEventListener('DOMNodeRemoved', this._boundOnChange, false);
 
       window.removeEventListener('resize', this._boundOnChange, false);
     });