+++ css-components/www/patterns/lib/onsen/js/angular-onsenui.js
@@ -1,4 +1,4 @@
-/*! angular-onsenui.js for onsenui - v2.0.0-rc.8 - 2016-06-02 */
+/*! angular-onsenui.js for onsenui - v2.0.0-rc.10 - 2016-06-09 */
 /* Simple JavaScript Inheritance
  * By John Resig http://ejohn.org/
  * MIT Licensed.