+++ CHANGELOG.md
@@ -6,6 +6,7 @@ v2.0.0-rc.11
 ----
  * ons-input: Fix `value` property bug for radio and checkbox.
  * ons-navigator: Fix [#1449](https://github.com/OnsenUI/OnsenUI/issues/1449).
+ * ons-popover: Fix [#1450](https://github.com/OnsenUI/OnsenUI/issues/1450).
 
 v2.0.0-rc.10
 ----