+++ CHANGELOG.md
@@ -5,6 +5,10 @@ CHANGELOG
 v2.0.0-rc.11
 ----
  * ons-input: Fix `value` property bug for radio and checkbox.
+ * ons-navigator: Fix [#1449](https://github.com/OnsenUI/OnsenUI/issues/1449).
+ * ons-popover: Fix [#1450](https://github.com/OnsenUI/OnsenUI/issues/1450).
+ * ons-navigator: Fix [#1389](https://github.com/OnsenUI/OnsenUI/issues/1389).
+ * ons-popover: Fix [#1388](https://github.com/OnsenUI/OnsenUI/issues/1388).
 
 v2.0.0-rc.10
 ----