+++ CHANGELOG.md
@@ -9,6 +9,7 @@ v2.0.0-rc.11
  * ons-popover: Fix [#1450](https://github.com/OnsenUI/OnsenUI/issues/1450).
  * ons-navigator: Fix [#1389](https://github.com/OnsenUI/OnsenUI/issues/1389).
  * ons-popover: Fix [#1388](https://github.com/OnsenUI/OnsenUI/issues/1388).
+ * ons-navigator: Fix [#1430](https://github.com/OnsenUI/OnsenUI/issues/1430).
  * ons-splitter-side: Use imported `orientation` object instead of global.
 
 v2.0.0-rc.10