+++ CHANGELOG.md
@@ -11,6 +11,7 @@ v2.0.0-rc.11
  * ons-popover: Fix [#1388](https://github.com/OnsenUI/OnsenUI/issues/1388).
  * ons-navigator: Fix [#1430](https://github.com/OnsenUI/OnsenUI/issues/1430).
  * ons-splitter-side: Use imported `orientation` object instead of global.
+ * ons-navigator: Fix [#1453](https://github.com/OnsenUI/OnsenUI/issues/1453).
 
 v2.0.0-rc.10
 ----